# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EquityLens API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/news/all": {
      "get": {
        "tags": ["importing news"],
        "summary": "Get News",
        "operationId": "get_news_api_news_all_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/news/": {
      "get": {
        "tags": ["importing news"],
        "summary": "Get News",
        "operationId": "get_news_api_news__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "business",
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/news/portfolio-tickers": {
      "get": {
        "tags": ["importing news"],
        "summary": "Get Portfolio Tickers",
        "operationId": "get_portfolio_tickers_api_news_portfolio_tickers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TickerResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/news/test-aapl/{ticker}": {
      "get": {
        "tags": ["importing news"],
        "summary": "Test Aapl News",
        "operationId": "test_aapl_news_api_news_test_aapl__ticker__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AAPLNewsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AAPLNewsResponse": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker",
            "examples": ["NPN"]
          },
          "total_articles": {
            "type": "integer",
            "title": "Total Articles",
            "examples": [3]
          },
          "positive": {
            "type": "integer",
            "title": "Positive",
            "examples": [1]
          },
          "negative": {
            "type": "integer",
            "title": "Negative",
            "examples": [0]
          },
          "neutral": {
            "type": "integer",
            "title": "Neutral",
            "examples": [2]
          },
          "articles": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Articles"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "total_articles",
          "positive",
          "negative",
          "neutral",
          "articles"
        ],
        "title": "AAPLNewsResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NewsResponse": {
        "properties": {
          "total_articles": {
            "type": "integer",
            "title": "Total Articles",
            "examples": [23]
          },
          "positive": {
            "type": "integer",
            "title": "Positive",
            "examples": [33]
          },
          "negative": {
            "type": "integer",
            "title": "Negative",
            "examples": [24]
          },
          "neutral": {
            "type": "integer",
            "title": "Neutral",
            "examples": [23]
          },
          "results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "total_articles",
          "positive",
          "negative",
          "neutral",
          "results"
        ],
        "title": "NewsResponse"
      },
...
      "TickerResponse": {
        "properties": {
          "tickers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tickers",
            "examples": [["AAPL", "MFST", "TSLA"]]
          }
        },
        "type": "object",
        "required": ["tickers"],
        "title": "TickerResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}